refact: Move non-host fields to Organization
