+++ test/server/graphql/v2/query/AccountQuery.test.js
@@ -154,7 +154,7 @@ describe('server/graphql/v2/query/AccountQuery', () => {
     it('is public for host accounts', async () => {
       const hostAdminUser = await fakeUser();
       const randomUser = await fakeUser();
-      const host = await fakeHost({ location: { address: 'PUBLIC!' }, admin: hostAdminUser });
+      const host = await fakeHost({ hasHosting: true, location: { address: 'PUBLIC!' }, admin: hostAdminUser });
       const resultUnauthenticated = await graphqlQueryV2(accountQuery, { slug: host.slug });
       const resultRandomUser = await graphqlQueryV2(accountQuery, { slug: host.slug }, randomUser);
       const resultHostAdmin = await graphqlQueryV2(accountQuery, { slug: host.slug }, hostAdminUser);
